move benchmarks to tests.
