+++ src/coprocessor/codec/mysql/json/mod.rs
@@ -97,10 +97,6 @@ pub fn json_object(kvs: Vec<Datum>) -> Result<Json> {
 
 #[cfg(test)]
 mod test {
-    use std::fs::File;
-    use std::io::prelude::*;
-    use std::io::BufReader;
-    use std::path::{Path, PathBuf};
     use super::*;
 
     #[test]
@@ -120,61 +116,5 @@ mod test {
         for (d, ep_json) in cases {
             assert_eq!(json_array(d).unwrap(), ep_json);
         }
-    }
-
-    #[test]
-    fn test_json_object() {
-        let cases = vec![
-            vec![Datum::I64(1)],
-            vec![
-                Datum::I64(1),
-                Datum::Bytes(b"sdf".to_vec()),
-                Datum::Null,
-                Datum::U64(2),
-            ],
-        ];
-        for d in cases {
-            assert!(json_object(d).is_err());
-        }
-
-        let cases = vec![
-            (
-                vec![
-                    Datum::I64(1),
-                    Datum::Bytes(b"sdf".to_vec()),
-                    Datum::Bytes(b"asd".to_vec()),
-                    Datum::Bytes(b"qwe".to_vec()),
-                    Datum::I64(2),
-                    Datum::Json(r#"{"3":4}"#.parse().unwrap()),
-                ],
-                r#"{"1":"sdf","2":{"3":4},"asd":"qwe"}"#.parse().unwrap(),
-            ),
-            (vec![], "{}".parse().unwrap()),
-        ];
-        for (d, ep_json) in cases {
-            assert_eq!(json_object(d).unwrap(), ep_json);
-        }
-    }
-
-    #[inline]
-    fn open_file_read_lines<P: AsRef<Path>>(filename: P) -> Vec<String> {
-        File::open(filename)
-            .map(BufReader::new)
-            .and_then(|reader| {
-                reader.lines().collect::<::std::io::Result<Vec<_>>>()
-            })
-            .unwrap()
-    }
-
-    pub fn load_test_jsons() -> Vec<String> {
-        let mut file = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-        file.push("src/coprocessor/codec/mysql/json/world_bank.json");
-        open_file_read_lines(file)
-    }
-
-    #[test]
-    fn test_load_test_jsons() {
-        let json_texts = load_test_jsons();
-        assert_eq!(json_texts.len(), 500);
     }
 }
