+++ tests/coprocessor/codec/mysql/json/world_bank.json
