+++ tests/coprocessor/codec/mysql/mod.rs
@@ -1,0 +1,1 @@
+mod json;
