+++ tests/coprocessor/mod.rs
@@ -11,4 +11,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod codec;
 mod test_select;
